requirements.txt
gradio
ultralytics
Pillow